+++ changelog.md
@@ -11,6 +11,7 @@ Codion Change Log
 - EntityEditorPanel.Config.transferFocusOnEnter() added along with Config.TRANSFER_FOCUS_ON_ENTER.
 - EntityEditorPanel.InputFocus moved to EntityEditPanel.
 - EntityEditorPanel.Config.modifiedWarning() moved to EntityEditPanel.Config.
+- EntityEditPanel.Config.editPanel() removed, unused.
 
 ## 0.18.68
 ### is.codion.common.reactive